+++ .github/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - build scripts
  - added `upload-artifact` to github actions
  - `STLLoader` to load ascii or binary stl files
+ - `Fullscreen` for `View3d`
  
 ### Changed